Remove scala.meta.io.{Multipath,Sourcepath,Fragment}
These classes are no longer used.
