+++ tests/jvm/src/test/scala/scala/meta/tests/io/MultipathSuite.scala
@@ -2,7 +2,6 @@ package scala.meta.tests
 package io
 
 import java.io.File
-import java.net.URI
 import org.scalatest.FunSuite
 import scala.meta._
 
@@ -14,31 +13,6 @@ class MultipathSuite extends FunSuite {
   files.foreach(file => {
     assert(new File(tmp, file).createNewFile())
   })
-  val multipath = Sourcepath(tmp.getAbsolutePath)
-
-  test("Multipath.deep") {
-    val obtained = multipath.deep.map(_.syntax)
-    val expected = files.map(file => Fragment(AbsolutePath(tmp), RelativePath(file))).map(_.syntax)
-    assert(obtained.sorted == expected.sorted)
-  }
-
-  test("Multipath.find") {
-    assert(multipath.find(RelativePath("a")).nonEmpty)
-  }
-
-  test("Multipath.relativize") {
-    assert(multipath.relativize(new URI("blah")).isEmpty)
-    files.foreach { file =>
-      assert(multipath.relativize(new File(tmp, file).toURI).get.toString == file)
-    }
-  }
-
-  test("Sourcepath.syntax") {
-    val fromPath = Sourcepath(List(AbsolutePath(tmp)))
-    val fromSyntax = Sourcepath(tmp.getAbsolutePath)
-    assert(fromPath == fromSyntax)
-    assert(fromPath.syntax == fromSyntax.syntax)
-  }
 
   test("Classpath.syntax") {
     val fromPath = Classpath(List(AbsolutePath(tmp)))
